+++ authority-portal-backend/authority-portal-db/src/main/resources/db/migration/V2__Schema_Extension.sql
@@ -1,0 +1,2 @@
+-- User
+alter type user_registration_status add value 'INVITED' before 'CREATED';