+++ authority-portal-backend/authority-portal-quarkus/src/main/kotlin/de/sovity/authorityportal/web/UiResourceImpl.kt
@@ -6,6 +6,7 @@ import de.sovity.authorityportal.api.model.ConnectorOverviewResult
 import de.sovity.authorityportal.api.model.CreateConnectorRequest
 import de.sovity.authorityportal.api.model.CreateOrganizationRequest
 import de.sovity.authorityportal.api.model.IdResponse
+import de.sovity.authorityportal.api.model.InviteParticipantUserRequest
 import de.sovity.authorityportal.api.model.OrganizationDetailResult
 import de.sovity.authorityportal.api.model.OrganizationOverviewResult
 import de.sovity.authorityportal.api.model.UserInfo
@@ -18,6 +19,7 @@ import de.sovity.authorityportal.web.pages.connectormanagement.ConnectorManageme
 import de.sovity.authorityportal.web.pages.organizationmanagement.OrganizationManagementApiService
 import de.sovity.authorityportal.web.pages.usermanagement.UserInfoApiService
 import de.sovity.authorityportal.web.pages.usermanagement.UserRoleApiService
+import de.sovity.authorityportal.web.pages.userregistration.UserInvitationApiService
 import de.sovity.authorityportal.web.pages.userregistration.UserRegistrationApiService
 import jakarta.inject.Inject
 import jakarta.transaction.Transactional
@@ -37,6 +39,9 @@ class UiResourceImpl : UiResource {
 
     @Inject
     lateinit var userRegistrationApiService: UserRegistrationApiService
+
+    @Inject
+    lateinit var userInvitationApiService: UserInvitationApiService
 
     @Inject
     lateinit var organizationManagementApiService: OrganizationManagementApiService
@@ -69,15 +74,22 @@ class UiResourceImpl : UiResource {
     @Transactional
     override fun changeParticipantRole(userId: String, roleDto: UserRoleDto): IdResponse {
         authUtils.requiresRole(Roles.UserRoles.PARTICIPANT_ADMIN)
-        authUtils.assertLoggedInUserInSameOrgAs(userId)
+        authUtils.requiresMemberOfSameOrganizationAs(userId)
         return userRoleApiService.changeParticipantRole(userId, roleDto, loggedInUser.organizationMdsId!!, loggedInUser.userId)
+    }
+
+    @Transactional
+    override fun inviteUser(invitationInformation: InviteParticipantUserRequest): IdResponse {
+        authUtils.requiresRole(Roles.UserRoles.PARTICIPANT_ADMIN)
+        authUtils.requiresMemberOfAnyOrganization()
+        return userInvitationApiService.inviteParticipantUser(invitationInformation, loggedInUser.organizationMdsId!!, loggedInUser.userId)
     }
 
     // Organization management (Authority)
     @Transactional
     override fun changeAuthorityRole(userId: String, roleDto: UserRoleDto): IdResponse {
         authUtils.requiresRole(Roles.UserRoles.AUTHORITY_ADMIN)
-        authUtils.assertLoggedInUserInSameOrgAs(userId)
+        authUtils.requiresMemberOfSameOrganizationAs(userId)
         return userRoleApiService.changeAuthorityRole(userId, roleDto, loggedInUser.userId)
     }
 
@@ -109,14 +121,14 @@ class UiResourceImpl : UiResource {
     @Transactional
     override fun ownOrganizationConnectors(): ConnectorOverviewResult {
         authUtils.requiresRole(Roles.UserRoles.PARTICIPANT_USER)
-        authUtils.requiresMemberOfOrganization()
+        authUtils.requiresMemberOfAnyOrganization()
         return connectorManagementApiService.listOwnOrganizationConnectors(loggedInUser.organizationMdsId!!, loggedInUser.userId)
     }
 
     @Transactional
     override fun ownOrganizationConnectorDetails(connectorId: String): ConnectorDetailDto {
         authUtils.requiresRole(Roles.UserRoles.PARTICIPANT_USER)
-        authUtils.requiresMemberOfOrganization()
+        authUtils.requiresMemberOfAnyOrganization()
         return connectorManagementApiService.ownOrganizationConnectorDetails(connectorId, loggedInUser.organizationMdsId!!, loggedInUser.userId)
     }
 
@@ -135,21 +147,21 @@ class UiResourceImpl : UiResource {
     @Transactional
     override fun createOwnConnector(connector: CreateConnectorRequest): IdResponse {
         authUtils.requiresRole(Roles.UserRoles.PARTICIPANT_CURATOR)
-        authUtils.requiresMemberOfOrganization()
+        authUtils.requiresMemberOfAnyOrganization()
         return connectorManagementApiService.createOwnConnector(connector, loggedInUser.organizationMdsId!!, loggedInUser.userId)
     }
 
     @Transactional
     override fun deleteOwnConnector(connectorId: String): IdResponse {
         authUtils.requiresRole(Roles.UserRoles.PARTICIPANT_CURATOR)
-        authUtils.requiresMemberOfOrganization()
+        authUtils.requiresMemberOfAnyOrganization()
         return connectorManagementApiService.deleteOwnConnector(connectorId, loggedInUser.organizationMdsId!!, loggedInUser.userId)
     }
 
     @Transactional
     override fun createProvidedConnector(mdsId: String, connector: CreateConnectorRequest): IdResponse {
         authUtils.requiresAnyRole(Roles.UserRoles.AUTHORITY_ADMIN, Roles.UserRoles.SERVICE_PARTNER_ADMIN)
-        authUtils.requiresMemberOfOrganization()
+        authUtils.requiresMemberOfAnyOrganization()
         return connectorManagementApiService.createProvidedConnector(connector, mdsId, loggedInUser.organizationMdsId!!, loggedInUser.userId)
     }