+++ authority-portal-backend/authority-portal-quarkus/src/main/kotlin/de/sovity/authorityportal/web/pages/userregistration/UserRegistrationStatusMapper.kt
@@ -5,6 +5,7 @@ import de.sovity.authorityportal.db.jooq.enums.UserRegistrationStatus
 
 fun UserRegistrationStatus.toDto(): UserRegistrationStatusDto = when (this) {
     UserRegistrationStatus.FIRST_USER -> UserRegistrationStatusDto.FIRST_USER
+    UserRegistrationStatus.INVITED -> UserRegistrationStatusDto.INVITED
     UserRegistrationStatus.CREATED -> UserRegistrationStatusDto.CREATED
     UserRegistrationStatus.PENDING -> UserRegistrationStatusDto.PENDING
     UserRegistrationStatus.APPROVED -> UserRegistrationStatusDto.APPROVED
@@ -13,6 +14,7 @@ fun UserRegistrationStatus.toDto(): UserRegistrationStatusDto = when (this) {
 
 fun UserRegistrationStatusDto.toDb(): UserRegistrationStatus = when (this) {
     UserRegistrationStatusDto.FIRST_USER -> UserRegistrationStatus.FIRST_USER
+    UserRegistrationStatusDto.INVITED -> UserRegistrationStatus.INVITED
     UserRegistrationStatusDto.CREATED -> UserRegistrationStatus.CREATED
     UserRegistrationStatusDto.PENDING -> UserRegistrationStatus.PENDING
     UserRegistrationStatusDto.APPROVED -> UserRegistrationStatus.APPROVED