+++ authority-portal-backend/authority-portal-quarkus/src/main/kotlin/de/sovity/authorityportal/web/services/FirstLoginService.kt
@@ -1,0 +1,16 @@
+package de.sovity.authorityportal.web.services
+
+import de.sovity.authorityportal.db.jooq.enums.UserRegistrationStatus
+import de.sovity.authorityportal.db.jooq.tables.records.UserRecord
+import jakarta.enterprise.context.ApplicationScoped
+
+@ApplicationScoped
+class FirstLoginService {
+
+    fun approveIfInvited(user: UserRecord) {
+        if (user.registrationStatus == UserRegistrationStatus.INVITED) {
+            user.registrationStatus = UserRegistrationStatus.APPROVED
+            user.update()
+        }
+    }
+}